+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.8.4](https://github.com/salesforcecli/plugin-source/compare/v1.8.3...v1.8.4) (2022-01-20)
+
+### [1.8.3](https://github.com/salesforcecli/plugin-source/compare/v1.8.2...v1.8.3) (2022-01-20)
+
+### [1.8.2](https://github.com/salesforcecli/plugin-source/compare/v1.8.1...v1.8.2) (2022-01-20)
+
+### Bug Fixes
+
+- include correct username for deploy:report in async deploy ([#392](https://github.com/salesforcecli/plugin-source/issues/392)) ([f58ff25](https://github.com/salesforcecli/plugin-source/commit/f58ff251c60b18391374de7a16f732540515d524))
+
+### [1.8.1](https://github.com/salesforcecli/plugin-source/compare/v1.8.0...v1.8.1) (2022-01-19)
+
+### Bug Fixes
+
+- missing quotes in --help for source:retrieve ([#393](https://github.com/salesforcecli/plugin-source/issues/393)) ([2c1e5c6](https://github.com/salesforcecli/plugin-source/commit/2c1e5c6157aed2c1d855af8f2c0a1ebf4478dc23))
+
 ## [1.8.0](https://github.com/salesforcecli/plugin-source/compare/v1.6.2...v1.8.0) (2022-01-13)
 
 ### Features